simplify tests in all samples
