--- samples/jvm/junit5/src/test/kotlin/ch/tutteli/SmokeTest.kt
@@ -1,66 +1,0 @@
-package ch.tutteli
-
-import org.junit.jupiter.api.*;
-
-import ch.tutteli.atrium.api.fluent.en_GB.*
-import ch.tutteli.atrium.api.verbs.assertThat
-import ch.tutteli.atrium.assertions.Assertion
-import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.domain.builders.AssertImpl
-import ch.tutteli.atrium.reporting.RawString
-import ch.tutteli.atrium.reporting.translating.StringBasedTranslatable
-import ch.tutteli.atrium.translations.DescriptionBasic
-
-class SmokeTest {
-    @Test
-    fun toBe_canBeUsed() {
-        assertThat(1).isLessThan(2)
-    }
-
-    @Test
-    fun assertionFunctionWithoutI18nCanBeUsed() {
-        assertThat(2).isEven()
-    }
-
-    @Test
-    fun assertionFunctionWithI18nCanBeUsed() {
-        assertThat(4).isMultipleOf(2)
-    }
-
-    @Test
-    fun assertAnExceptionOccurred() {
-        assertThat {
-            throw IllegalArgumentException()
-        }.toThrow<IllegalArgumentException>()
-    }
-
-    @Test
-    fun assertAnExceptionWithAMessageOccurred() {
-        assertThat {
-            throw IllegalArgumentException("oho... hello btw")
-        }.toThrow<IllegalArgumentException>{
-            messageContains("hello")
-        }
-    }
-
-    @Test
-    fun assertNotToThrow() {
-        assertThat {
-
-        }.notToThrow()
-    }
-}
-
-fun Expect<Int>.isEven() =
-    createAndAddAssertion(DescriptionBasic.IS, RawString.create("an even number")) { it % 2 == 0 }
-
-fun Expect<Int>.isMultipleOf(base: Int) = addAssertion(_isMultipleOf(this, base))
-
-fun _isMultipleOf(expect: Expect<Int>, base: Int): Assertion =
-    AssertImpl.builder.createDescriptive(expect, DescriptionIntAssertions.IS_MULTIPLE_OF, base) {
-        it % base == 0
-    }
-
-enum class DescriptionIntAssertions(override val value: String) : StringBasedTranslatable {
-    IS_MULTIPLE_OF("is multiple of")
-}
